Abstract common test functions to own file
